fix: deserializing errors for recorded API response
