+++ src/main/java/systems/boos/traindelays/TimetableStop.java
@@ -1,44 +1,27 @@
 package systems.boos.traindelays;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper;
 import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlProperty;
 
-import java.time.Instant;
-import java.util.Objects;
+import java.util.ArrayList;
+import java.util.List;
 
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class TimetableStop {
 
-    @JacksonXmlProperty(isAttribute = true, localName = "ct")
-    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyMMddHHmm", timezone = "Europe/Berlin")
-    private Instant changedTime;
+    @JacksonXmlElementWrapper(useWrapping = false)
+    @JacksonXmlProperty(localName = "dp")
+    private List<Event> departures = new ArrayList<>();
 
-    public Instant getChangedTime() {
-        return changedTime;
+    @JacksonXmlProperty(isAttribute = true)
+    private String id;
+
+    public List<Event> getDepartures() {
+        return departures;
     }
 
-    public void setChangedTime(Instant changedTime) {
-        this.changedTime = changedTime;
-    }
-
-    @Override
-    public String toString() {
-        return "TimetableStop{" +
-                "changedTime='" + changedTime + '\'' +
-                '}';
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-
-        TimetableStop that = (TimetableStop) o;
-
-        return Objects.equals(changedTime, that.changedTime);
-    }
-
-    @Override
-    public int hashCode() {
-        return changedTime != null ? changedTime.hashCode() : 0;
+    public void setDepartures(List<Event> departures) {
+        this.departures = departures;
     }
 }
