+++ src/test/java/systems/boos/traindelays/unit/DeserializeTimetableApiResponseBodyTest.java
@@ -2,6 +2,7 @@ package systems.boos.traindelays.unit;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.dataformat.xml.XmlMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import systems.boos.traindelays.Timetable;
@@ -11,6 +12,7 @@ class DeserializeTimetableApiResponseBodyTest {
     @Test
     void DeserializeTimetableApiResponseBody() throws JsonProcessingException {
         XmlMapper mapper = new XmlMapper();
+        mapper.registerModule(new JavaTimeModule());
 
         String xmlAsString = TimetableApiResponses.getRecordedResponse();
